Use REG ADD if the script is elevated
The method of creating and invoking a .reg file is beneficial in that
the script doesn't have to be elevated, but it isn't seamless. Invoking
the .reg file launches the UAC prompt and displays a warning dialog
about modifying the registry. If the script is elevated, we can avoid
that by directly invoking REG ADD.

This commit also adds verbose and quiet flags, and adds more details to
the readme.