Add option to refresh metadata
Allow the user to add an additional 'Refresh Plex Metadata' context menu
entry that attempts to refresh metadata for all files that live in the
given directory. It does this by querying for all movies/episodes/albums
in a library and checking whether it has a file that lives within the
given directory. Note that this is much less efficient than the scan
process, and is disabled by default.