Ignore unset store_true arguments
get_config_value tries to detect whether a given argument is provided in
both the config file and the command line, with the command line value
taking precedence. However, that check is done incorrectly, as
store_true values like --web will be set to False if not present, while
the script only checks for None. Also ignore False.